requirements.txt
absl-py==0.14.0
argon2-cffi==21.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
contextlib2==21.6.0
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
deepface==0.0.68
defusedxml==0.7.1
Django==3.2.7
easyocr==1.4.1
entrypoints==0.3
executing==0.8.2
filelock==3.1.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gdown==3.14.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
idna==3.2
image==1.5.33
image-quality==1.2.7
imageio==2.9.0
importlib-metadata==4.10.1
importlib-resources==5.4.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libsvm==3.23.0.4
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mediapipe==0.8.9.1
mistune==0.8.4
mtcnn==0.1.1
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
object-detection==0.0.3
object-detection-0.1==0.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
Pillow-PIL==0.1.dev0
platformdirs==2.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.1.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.8.30
tomli==2.0.0
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.7.0
